+++ src/index.ts
@@ -21,6 +21,7 @@ export {
   type CopyCloneMode,
   type RootCreateJsonOptions,
   type RootCreateOptions,
+  type RootCreateStreamOptions,
   type RootDefaults,
   type RootMkdirOptions,
   type RootMoveOptions,